+++ include/NodeEngine/detail/TupleBufferImpl.hpp
@@ -79,13 +79,13 @@ class BufferControlBlock {
      * @brief method to get the watermark as a timestamp
      * @return watermark
      */
-    size_t getWatermark() const;
+    int64_t getWatermark() const;
 
     /**
    * @brief method to set the watermark with a timestamp
    * @param value timestamp
    */
-    void setWatermark(size_t value);
+    void setWatermark(int64_t value);
 
 #ifdef NES_DEBUG_TUPLE_BUFFER_LEAKS
     void dumpOwningThreadInfo();
@@ -94,7 +94,7 @@ class BufferControlBlock {
   private:
     std::atomic<uint32_t> referenceCounter;
     std::atomic<uint32_t> numberOfTuples;
-    std::atomic<size_t>  watermark;
+    std::atomic<int64_t>  watermark;
     MemorySegment* owner;
     std::function<void(MemorySegment*)> recycleCallback;
 #ifdef NES_DEBUG_TUPLE_BUFFER_LEAKS
