+++ src/NodeEngine/TupleBuffer.cpp
@@ -96,12 +96,12 @@ void TupleBuffer::revertEndianness(SchemaPtr schema) {
     detail::revertEndianness(*this, schema);
 }
 
-size_t TupleBuffer::getWatermark()
+int64_t TupleBuffer::getWatermark()
 {
     return controlBlock->getWatermark();
 }
 
-void TupleBuffer::setWatermark(size_t value)
+void TupleBuffer::setWatermark(int64_t value)
 {
     controlBlock->setWatermark(value);
 }
